% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{comp_correlation}
\alias{comp_correlation}
\title{Calculate correlation coefficients between two compartment tracks}
\usage{
comp_correlation(x, y, correlation = c("pearson", "spearman", "kendall"))
}
\arguments{
\item{x}{Compartment scores. Should be of bedtorch format.}

\item{y}{Compartment scores.}
}
\value{
A named list of correlation coefficients
}
\description{
Calculate correlation coefficients between two compartment tracks
}
