% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{compartment_fanc}
\alias{compartment_fanc}
\title{Get compartment using FANC, from an existing .hic file}
\usage{
compartment_fanc(
  hic_file,
  norm = c("NONE", "VC", "VC_SQRT", "KR", "SCALE"),
  resol,
  ev = 1:2,
  genome = NULL,
  ab_file = NULL,
  keep_ab = !is_null(ab_file)
)
}
\arguments{
\item{ev}{an integer vector specifying which eigenvectors to calculate}

\item{ab_file}{path to a pre-calculated AB matrix file}

\item{keep_ab}{if \code{TRUE}, keep the AB matrix file. This requires \code{ab_file = NULL}}
}
\description{
Get compartment using FANC, from an existing .hic file
}
