% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{compartment_ht}
\alias{compartment_ht}
\title{Compute compartment scores using hictools's own implementation of compartment analysis, as opposed to juicer}
\usage{
compartment_ht(
  hic_matrix,
  method = c("lieberman", "obs_exp", "nonzero", "average"),
  oe = c("ht", "as-is"),
  chrom = NULL,
  npc = 2L
)
}
\arguments{
\item{npc}{A positive integer indicating how many principal components to keep.}
}
\description{
Compute compartment scores using hictools's own implementation of compartment analysis, as opposed to juicer
}
