% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{compartment_juicer}
\alias{compartment_juicer}
\title{Calculate compartment scores using Juicer tools}
\usage{
compartment_juicer(
  hic_matrix,
  juicertools,
  java = "java",
  norm = "NONE",
  resol = NULL
)
}
\description{
Calculate compartment scores using Juicer tools
}
