% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{convert_hic_matrix}
\alias{convert_hic_matrix}
\title{Convert a Hi-C map to a naive matrix}
\usage{
convert_hic_matrix(hic_matrix, chrom = NULL, missing_score = NA)
}
\arguments{
\item{missing_score}{Scores where the bin does not contain any signal.}
}
\description{
Convert a Hi-C map to a naive matrix
}
