% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{decay}
\alias{decay}
\title{Apply a "decay curve" to a Hi-C matrix}
\usage{
decay(hic_matrix, decay_curve)
}
\arguments{
\item{decay_curve}{A data frame with two columns: \code{dist} and \code{factor}}
}
\description{
This may be useful when \code{hic_matrix} is a cofrag dataset
}
