% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{dump_juicer_hic}
\alias{dump_juicer_hic}
\title{Dump a Hi-C object as .hic format}
\usage{
dump_juicer_hic(
  hic_matrix,
  file_path,
  juicertools,
  java = "java",
  ref_genome = "hg19"
)
}
\arguments{
\item{hic_matrix}{A Hi-C object}

\item{file_path}{Path to the output .hic file}

\item{java}{Path to JVM. Default is \code{java}}

\item{ref_genome}{Reference genome \code{hic_matrix} is using. Default is \code{hg19}}
}
\description{
This function invokes Juicer tools for the dumping
}
