% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{flip_compartment}
\alias{flip_compartment}
\title{Flip the compartment score signs according to \code{standard}}
\usage{
flip_compartment(compartment, standard, resol)
}
\arguments{
\item{standard}{Can be either \code{gene_density.hg19} or a BED-format data frame.
If \code{gene_density.hg19}, the function will load the gene density data.
Currently only supports a number of resolutions: 50k, 100k, 200k, 500k, 1m,
2.5m.}
}
\description{
Flip the compartment score signs according to \code{standard}
}
