% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gc.GRCh37.100kbp}
\alias{gc.GRCh37.100kbp}
\alias{gc.GRCh37.250kbp}
\alias{gc.GRCh37.500kbp}
\alias{gc.GRCh37.1000kbp}
\alias{gc.GRCh37.2500kbp}
\alias{gc.GRCh38.100kbp}
\alias{gc.GRCh38.250kbp}
\alias{gc.GRCh38.500kbp}
\alias{gc.GRCh38.1000kbp}
\alias{gc.GRCh38.2500kbp}
\title{GC content}
\format{
An object of class \code{GRanges} of length 28529.

An object of class \code{GRanges} of length 11440.

An object of class \code{GRanges} of length 5738.

An object of class \code{GRanges} of length 2888.

An object of class \code{GRanges} of length 1174.

An object of class \code{GRanges} of length 29237.

An object of class \code{GRanges} of length 11709.

An object of class \code{GRanges} of length 5862.

An object of class \code{GRanges} of length 2941.

An object of class \code{GRanges} of length 1188.
}
\usage{
gc.GRCh37.100kbp

gc.GRCh37.250kbp

gc.GRCh37.500kbp

gc.GRCh37.1000kbp

gc.GRCh37.2500kbp

gc.GRCh38.100kbp

gc.GRCh38.250kbp

gc.GRCh38.500kbp

gc.GRCh38.1000kbp

gc.GRCh38.2500kbp
}
\description{
Fixed-interval GC content for a number of reference genomes, with different bin sizes.
}
\keyword{datasets}
