% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gencode.v30.b37.gene_density.50kbp}
\alias{gencode.v30.b37.gene_density.50kbp}
\alias{gencode.v30.b37.gene_density.100kbp}
\alias{gencode.v30.b37.gene_density.250kbp}
\alias{gencode.v30.b37.gene_density.500kbp}
\alias{gencode.v30.b37.gene_density.1000kbp}
\title{Gene density}
\format{
An object of class \code{GRanges} of length 62072.

An object of class \code{GRanges} of length 31061.

An object of class \code{GRanges} of length 12459.

An object of class \code{GRanges} of length 6267.

An object of class \code{GRanges} of length 3173.

An object of class \code{GRanges} of length 12459.
}
\usage{
gencode.v30.b37.gene_density.50kbp

gencode.v30.b37.gene_density.100kbp

gencode.v30.b37.gene_density.250kbp

gencode.v30.b37.gene_density.500kbp

gencode.v30.b37.gene_density.1000kbp

gencode.v30.b37.gene_density.250kbp
}
\description{
Gene density data from \code{gencode.v30.b37.gene_density}. For a given interval,
indicate the number of genes overlapping with that interval.
}
\details{
The dataset comes with various interval widths: 50 kbp, 100 kbp, 250 kbp, 500
kbp, 1000 kbp and 2500 kbp.

To prepare the dataset:

\verb{bedtools makewindows -g human_g1k_v37.chrom.sizes -w 500000 |}
\verb{bedtools map -b gencode.v30.b37.gene_density.bedgraph -a - -c 4 -o count}
}
\examples{

data(gencode.v30.b37.gene_density.50kbp)
}
\keyword{datasets}
