% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_compartment.character}
\alias{get_compartment.character}
\title{Calculate compartment scores}
\usage{
\method{get_compartment}{character}(
  hic_matrix,
  method = c("juicer", "lieberman", "obs_exp", "nonzero", "average", "fanc"),
  chrom = NULL,
  resol,
  standard = NULL,
  oe = c("juicer", "ht", "as-is"),
  genome = NULL,
  norm = c("NONE", "VC", "VC_SQRT", "KR", "SCALE"),
  ...
)
}
\arguments{
\item{resol}{A positive integer for resolution.}

\item{genome}{Name of the reference genome.}
}
\description{
Load Hi-C matrix from .hic file and calculate compartment scores.
}
