% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_compartment.ht_table}
\alias{get_compartment.ht_table}
\title{Calculate compartment scores}
\usage{
\method{get_compartment}{ht_table}(
  hic_matrix,
  method = c("juicer", "lieberman", "obs_exp", "nonzero", "average", "fanc"),
  chrom = NULL,
  reference = NULL,
  oe = c("juicer", "ht", "as-is"),
  juicertools = get_juicer_tools(),
  java = "java",
  ...
)
}
\arguments{
\item{hic_matrix}{A \code{ht_table} object for the Hi-C matrix.}
}
\description{
Calculate compartment scores for Hi-C matrix in ht_table format.
}
