% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_juicer_tools}
\alias{get_juicer_tools}
\title{Return the path of juicer_tools which is shipped along with hictools}
\usage{
get_juicer_tools()
}
\description{
Return the path of juicer_tools which is shipped along with hictools
}
