% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{hicexplorer_pca}
\alias{hicexplorer_pca}
\title{Run hicPCA}
\usage{
hicexplorer_pca(
  hic_matrix,
  juicertools,
  executable = "hicPCA",
  executable_convert = "hicConvertFormat",
  java = "java",
  npc = 2,
  chrom = NULL,
  method = "lieberman"
)
}
\arguments{
\item{hic_matrix}{A Hi-C matrix object}
}
\description{
Run hicPCA and return the PCs, the O/E matrix and the Pearson correlation matrix
}
