% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_resol}
\alias{is_valid_resol}
\title{Is the resolution valid? Only support a limited number of resolution choices}
\usage{
is_valid_resol(resol)
}
\description{
Is the resolution valid? Only support a limited number of resolution choices
}
