% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_wholenumber}
\alias{is_wholenumber}
\title{Test if \code{x} is a vector of whole numbers}
\usage{
is_wholenumber(x, tol = .Machine$double.eps^0.5)
}
\description{
Test if \code{x} is a vector of whole numbers
}
