% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_hic}
\alias{load_hic}
\title{Load Hi-C dataset from file}
\usage{
load_hic(
  file_path,
  format = c("juicer_short", "juicer_dump", "juicer_hic", "genbed", "cool"),
  resol = NULL,
  ...
)
}
\description{
Load Hi-C dataset from file
}
