% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_hic_cool}
\alias{load_hic_cool}
\title{Load Hi-C data in cool format}
\usage{
load_hic_cool(
  file_path,
  chrom = NULL,
  matrix = "unknown",
  norm = "unknown",
  hdf5 = TRUE,
  cooler = "cooler"
)
}
\description{
Load Hi-C data in cool format
}
