% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{normalize_compartment}
\alias{normalize_compartment}
\title{Normalize the compartment score signs according to \code{standard}}
\usage{
normalize_compartment(compartment, standard, score_cols = c("score"))
}
\arguments{
\item{compartment}{A \code{GRanges} or \code{data.frame} containing compartment scores.}

\item{standard}{A \code{GRanges} or \code{data.frame} of standard compartment scores.}
}
\value{
The compartment. The signs of compartment scores may have been
flipped based on \code{standard}.
}
\description{
Normalize the compartment score signs according to \code{standard}
}
