% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{normalize_hic}
\alias{normalize_hic}
\title{Normalize Hi-C matrix}
\usage{
normalize_hic(hic_matrix, norm = "KR")
}
\arguments{
\item{norm}{Currently only KR is supported}
}
\description{
Normalize Hi-C matrix
}
