% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{oe_juicer}
\alias{oe_juicer}
\title{Calculate oe normalized matrix using Juicer tools}
\usage{
oe_juicer(
  hic_matrix,
  juicertools,
  java = "java",
  norm = "NONE",
  ref_genome = "hg19"
)
}
\arguments{
\item{hic_matrix}{A Hi-C object}
}
\description{
Calculate oe normalized matrix using Juicer tools
}
