% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{orient_compartment}
\alias{orient_compartment}
\title{Orient the compartment score signs according to \code{standard}}
\usage{
orient_compartment(compartment, standard, score_cols = c("score"))
}
\arguments{
\item{compartment}{A \code{GRanges} object for compartment scores.}

\item{standard}{A \code{GRanges} for the standard compartment track.}

\item{score_cols}{A character vector indicating which columns are candidates
for compartment scores. For example, \code{c("PC1", "PC2")} means from \code{PC1} and
\code{PC2}, picking the one most likely to be associated with A/B compartments,
and orient it if necessary.}
}
\value{
The compartment. The signs of compartment scores may have been
flipped based on \code{standard}.
}
\description{
Orient the compartment score signs according to \code{standard}
}
