% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{pearson_juicer}
\alias{pearson_juicer}
\title{Calculate O/E Pearson correlation matrix using Juicer tools}
\usage{
pearson_juicer(
  hic_matrix,
  chrom = NULL,
  juicertools = get_juicer_tools(),
  java = "java",
  norm = "NONE"
)
}
\description{
Calculate O/E Pearson correlation matrix using Juicer tools
}
