% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_compartment_scatter}
\alias{plot_compartment_scatter}
\title{Get the scatter plot between two compartment tracks}
\usage{
plot_compartment_scatter(
  x,
  y,
  size = 0.1,
  alpha = 0.5,
  add = "reg.line",
  add.params = list(fill = "lightgray"),
  conf.int = TRUE,
  stat_cor_method = c("pearson", "spearman", "kendall"),
  ...
)
}
\description{
Get the scatter plot between two compartment tracks
}
