% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{strat_dist}
\alias{strat_dist}
\title{Calculate distance-stratified summary}
\usage{
strat_dist(hic_matrix, smoothing = TRUE, min_nonzero = 10)
}
\arguments{
\item{hic_matrix}{A Hi-C object}

\item{smoothing}{Whether to apply smoothing for large-distance strata.
Default is \code{TRUE}.}

\item{min_nonzero}{If smoothing, in each stratum, the minimal number of
non-zero entries. Default is 10.}
}
\description{
Calculate distance-stratified summary
}
