% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_cool}
\alias{write_cool}
\title{Dump Hi-C data as .cool}
\usage{
write_cool(
  hic_matrix,
  file_path,
  juicertools,
  java = "java",
  executable = "hicConvertFormat"
)
}
\description{
Dump Hi-C data as .cool
}
