% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_hic_bedtorch}
\alias{write_hic_bedtorch}
\title{Write the Hi-C dataset to disk in bedtorch table format}
\usage{
write_hic_bedtorch(hic_matrix, file_path, comments = NULL)
}
\description{
Write the Hi-C dataset to disk in bedtorch table format
}
