% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_juicer_hic}
\alias{write_juicer_hic}
\title{Dump a Hi-C object as .hic format}
\usage{
write_juicer_hic(
  hic_matrix,
  file_path,
  juicertools = get_juicer_tools(),
  java = "java",
  ref_genome = c("hg19", "hg38"),
  norm = c("VC", "VC_SQRT", "KR", "SCALE")
)
}
\arguments{
\item{hic_matrix}{A Hi-C object}

\item{file_path}{Path to the output .hic file}

\item{java}{Path to JVM. Default is \code{java}}

\item{ref_genome}{Reference genome \code{hic_matrix} is using. Default is \code{hg19}}

\item{norm}{Calculate specific normalizations. Default is VC,VC_SQRT,KR,SCALE}
}
\description{
This function invokes Juicer tools for the dumping
}
