% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{comp_correlation}
\alias{comp_correlation}
\title{Calculate correlation coefficiencies between two compartment tracks}
\usage{
comp_correlation(x, y, correlation = c("pearson", "spearman", "kendall"))
}
\arguments{
\item{x}{Compartment scores.}

\item{y}{Compartment scores.}

\item{smooth}{A vector of positive integers. If \code{NULL}, will not apply smoothing.}
}
\value{
A named list of correlation coefficients
}
\description{
Calculate correlation coefficiencies between two compartment tracks
}
