% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{comp_correlation}
\alias{comp_correlation}
\title{Calculate correlation coefficients between two compartment tracks}
\usage{
comp_correlation(x, y, method = c("pearson", "spearman", "kendall"))
}
\arguments{
\item{x}{Compartment scores.}

\item{y}{Compartment scores.}

\item{method}{Type of correlation scores. Must be one of \code{"pearson"},
\code{"spearman"}, and \code{"kendall"}.}
}
\value{
A data frame for correlation.
}
\description{
Calculate correlation coefficients between two compartment tracks
}
