% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{compartment_fanc}
\alias{compartment_fanc}
\title{Get compartment using FANC, from an existing .hic file}
\usage{
compartment_fanc(
  hic_file,
  norm = c("NONE", "VC", "VC_SQRT", "KR", "SCALE"),
  resol,
  ev = 1:2,
  genome = c("hs37-1kg", "GRCh37", "GRCh38")
)
}
\description{
Get compartment using FANC, from an existing .hic file
}
