% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{compartment_juicer_file}
\alias{compartment_juicer_file}
\title{Get compartment using Juicer tools, from an existing .hic file}
\usage{
compartment_juicer_file(
  hic_file,
  juicertools = get_juicer_tools(),
  java = "java",
  norm = c("NONE", "VC", "VC_SQRT", "KR", "SCALE"),
  chrom = NULL,
  resol = NULL,
  standard = NULL,
  smoothing = NULL,
  genome = NULL
)
}
\description{
Get compartment using Juicer tools, from an existing .hic file
}
