% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{flip_compartment}
\alias{flip_compartment}
\title{Flip the compartment score signs according to \code{standard}}
\usage{
flip_compartment(compartment, standard)
}
\arguments{
\item{compartment}{A \code{GRanges} or \code{bedtorch_table} containing compartment scores.}

\item{standard}{A \code{GRanges} or \code{bedtorch_table} of standard compartment scores.}
}
\value{
The compartment. The signs of compartment scores may have been
flipped based on \code{standard}.
}
\description{
Flip the compartment score signs according to \code{standard}
}
