% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht_table.R
\name{ht_table}
\alias{ht_table}
\title{Construct an \code{ht_table} object, which represents a Hi-C-like dataset.}
\usage{
ht_table(
  dt,
  resol,
  type = c("observed", "oe", "cofrag"),
  norm = c("NONE", "KR", "VC", "VC_SQRT"),
  genome = NULL
)
}
\arguments{
\item{dt}{A \code{data.frame} input.}

\item{resol}{A positive integer for the resolution (or bin size).}

\item{type}{\code{observed} for raw counts, \code{oe} for "observation/expected" value,
and \code{cofrag} for cofragmentation contact score.}

\item{norm}{Indicate whether any normalization method has been applied.}

\item{genome}{A character scalar specifying the genome name. For example:
\verb{hg37-1kg}. Can be \code{NULL} if no genome is specified.}
}
\description{
Construct an \code{ht_table} object, which represents a Hi-C-like dataset.
}
