% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_hic_genbed}
\alias{load_hic_genbed}
\title{Load Hi-C data in BED format}
\usage{
load_hic_genbed(
  file_path,
  resol = NULL,
  chrom = NULL,
  type = c("observed", "oe", "cofrag"),
  norm = c("NONE", "KR", "VC", "VC_SQRT"),
  score_col = NULL,
  scale_score = TRUE,
  genome = NULL
)
}
\arguments{
\item{score_col}{Specify which column represents cofrag scores. Default is NULL,}

\item{scale_score}{Scale scales to the interval \link{0, 1}
which indicates the 7th column}
}
\description{
Load Hi-C data in BED format
}
