% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_juicer_hic}
\alias{load_juicer_hic}
\title{Load Hi-C dataset from a .hic file}
\usage{
load_juicer_hic(file_path, chrom, resol, matrix = "observed", norm = "NONE")
}
\arguments{
\item{file_path}{Path to the .hic file}

\item{chrom}{A character vector to indicate what chromosomes to load}

\item{resol}{Resolution in base pair (BP)}

\item{matrix}{Should be \code{observed} or \code{oe}}

\item{norm}{See the manual of Juicer tools.}
}
\value{
A Hi-C object
}
\description{
This function invokes Juicer tools for the dumping
}
