% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_compartment}
\alias{plot_compartment}
\title{Plot the A/B compartment profile}
\usage{
plot_compartment(comps, chrom = NULL, full_scale = FALSE)
}
\value{
A ggplot2 object showing the compartments
}
\description{
Plot the A/B compartment profile
}
