% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_hic_matrix}
\alias{plot_hic_matrix}
\title{Visualize Hi-C data}
\usage{
plot_hic_matrix(
  hic,
  chrom = NULL,
  control_hic = NULL,
  scale_factor = c(1, 1),
  norm = c(0, 1),
  transform = c("log10", "linear"),
  missing_value = NULL,
  n.breaks = NULL,
  color_palette = "viridis",
  gamma = "auto",
  tile_outline = NULL,
  verbose = FALSE
)
}
\arguments{
\item{chrom}{Indicate which chromosome to process. If \code{NULL}, \code{hic}
should contain only one chromosome, which will be used in the
visualization.}
}
\description{
Show a heatmap of the Hi-C data
}
