% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_hic_matrix}
\alias{plot_hic_matrix}
\title{Visualize Hi-C data}
\usage{
plot_hic_matrix(
  hic_matrix,
  chrom = NULL,
  control_hic_matrix = NULL,
  scale_factor = c(1, 1),
  norm = c(0, 1),
  transform = NULL,
  missing_value = NULL,
  n.breaks = NULL,
  color_palette = "viridis",
  matrix = "observed",
  gamma = 2.3,
  tile_outline = NULL
)
}
\arguments{
\item{chrom}{Indicate which chromosome to process. If \code{NULL}, \code{hic_matrix}
should contain only one chromosome, which will be used in the
visualization.}
}
\description{
Show a heatmap of the Hi-C data
}
